sort function against text contents is done.

+++ Assets/Autoya/Information/Tokenizer.cs
@@ -33,11 +33,11 @@ namespace AutoyaFramework.Information {
 
     public struct ContentAndWidthAndHeight {
 		public string content;
-		public float width;
-        public float totalHeight;
-        public ContentAndWidthAndHeight (string content, float width, float totalHeight) {
+        public float width;
+        public int totalHeight;
+        public ContentAndWidthAndHeight (string content, float width, int totalHeight) {
 			this.content = content;
-			this.width = width;
+            this.width = width;
             this.totalHeight = totalHeight;
         }
     }
